cleaning server code
