+++ client/src/redux/api/index.js
@@ -1,5 +1,5 @@
 export const fetchCommits = async (sha, page) => {
-  const response = await fetch(`/api/commits?sha=${sha}&page=${page}`);
+  const response = await fetch(`/api/github/commits?sha=${sha}&page=${page}`);
   const data = await response.json();
   if (response.status >= 400) {
     throw new Error(data.errors);
@@ -12,7 +12,7 @@ export const fetchTests = async (startrev, endrev, page) => {
   let keepGoing = true;
   while (keepGoing) {
     const response = await fetch(
-      `/api/testman?startrev=${startrev}&endrev=${endrev}&page=${page}`
+      `/api/testman/testman?startrev=${startrev}&endrev=${endrev}&page=${page}`
     );
     let data = await response.json();
     results.push(data.results.result);
@@ -30,7 +30,7 @@ export const fetchTests = async (startrev, endrev, page) => {
 
 export const fetchBuildSets = async str => {
   if (str) {
-    const response = await fetch(`/api/buildsets?${str}`);
+    const response = await fetch(`/api/buildbot/buildsets?${str}`);
     const data = await response.json();
     if (response.status >= 400) {
       throw new Error(data.errors);
@@ -41,7 +41,7 @@ export const fetchBuildSets = async str => {
 
 export const fetchBuildReq = async str => {
   if (str) {
-    const response = await fetch(`/api/buildreq?${str}`);
+    const response = await fetch(`/api/buildbot/buildreq?${str}`);
     const data = await response.json();
     if (response.status >= 400) {
       throw new Error(data.errors);
@@ -51,7 +51,7 @@ export const fetchBuildReq = async str => {
 };
 
 export const fetchBuilds = async str => {
-  const response = await fetch(`/api/builds?${str}`);
+  const response = await fetch(`/api/buildbot/builds?${str}`);
   const data = await response.json();
   if (response.status >= 400) {
     throw new Error(data.errors);
@@ -60,7 +60,7 @@ export const fetchBuilds = async str => {
 };
 
 export const fetchBuilders = async () => {
-  const response = await fetch('/api/builders');
+  const response = await fetch('/api/buildbot/builders');
   const data = await response.json();
   if (response.status >= 400) {
     throw new Error(data.errors);
@@ -69,7 +69,7 @@ export const fetchBuilders = async () => {
 };
 
 export const fetchBranches = async () => {
-  const response = await fetch('/api/branches');
+  const response = await fetch('/api/github/branches');
   const data = await response.json();
   if (response.status >= 400) {
     throw new Error(data.errors);
@@ -77,7 +77,7 @@ export const fetchBranches = async () => {
   return data;
 };
 export const fetchPulls = async (state, page) => {
-  const response = await fetch(`/api/pulls?state=${state}&page=${page}`);
+  const response = await fetch(`/api/github/pulls?state=${state}&page=${page}`);
   const data = await response.json();
   if (response.status >= 400) {
     throw new Error(data.errors);
